// transtemare-react/src/types/carpeta.ts
export interface CarpetaFormData {
  idCarpeta: string
  referenciaDestino: string
  nroTransmision: string
  numeroContenedorParte1: string
  numeroContenedorParte2: string
  numeroContenedorParte3: string
  transitoAduanero: boolean
  terminalId: string
  ciudadDeOrigen: string
  aduanaOrigen: string
  ciudadDeDestino: string
  aduanaDestino: string
  ciudadDeEmision: string
  ciudadEntrega: string
  paisOrigenMercaderias: string
  ruta: string
  cliente: string
  empresaRemitente: string
  empresaDestinataria: string
  empresaConsignataria: string
  notificarA: string
  despachante: string
  agenciaMaritima: string
  transportadoraCamionOriginal: string
  choferOriginal: string
  camionOriginal: string
  remolqueOriginal: string
  transportadoraCamionSustituto: string
  choferSubstituto: string
  camionSubstituto: string
  remolqueSubstituto: string
  tipoBulto: string
  tipoContenedor: string
  cantidadBultos: string
  pesoBruto: string
  volumenMC: string
  moneda: string
  valorFOT: string
  costoFlete: string
  seguro: string
  documentosAnexos: string
  formalidadesAduana: string
  rutasLargo: string
  rutasCorto: string
  validoHasta: string
  firmante: string
  cargarInformacionGarantia: boolean
  tipoGarantia: string
  importeGarantia: string
  bancoGarantia: string
  nroChequeGarantia: string
}

export const DEFAULT_CARPETA_FORM: CarpetaFormData = {
  idCarpeta: '',
  referenciaDestino: '',
  nroTransmision: '',
  numeroContenedorParte1: '',
  numeroContenedorParte2: '',
  numeroContenedorParte3: '',
  transitoAduanero: false,
  terminalId: '',
  ciudadDeOrigen: '',
  aduanaOrigen: '',
  ciudadDeDestino: '',
  aduanaDestino: '',
  ciudadDeEmision: '',
  ciudadEntrega: '',
  paisOrigenMercaderias: '',
  ruta: '',
  cliente: '',
  empresaRemitente: '',
  empresaDestinataria: '',
  empresaConsignataria: '',
  notificarA: '',
  despachante: '',
  agenciaMaritima: '',
  transportadoraCamionOriginal: '',
  choferOriginal: '',
  camionOriginal: '',
  remolqueOriginal: '',
  transportadoraCamionSustituto: '',
  choferSubstituto: '',
  camionSubstituto: '',
  remolqueSubstituto: '',
  tipoBulto: '',
  tipoContenedor: '',
  cantidadBultos: '',
  pesoBruto: '',
  volumenMC: '',
  moneda: 'US$',
  valorFOT: '',
  costoFlete: '',
  seguro: '',
  documentosAnexos: '',
  formalidadesAduana: '',
  rutasLargo: '',
  rutasCorto: '',
  validoHasta: '',
  firmante: '',
  cargarInformacionGarantia: false,
  tipoGarantia: '',
  importeGarantia: '',
  bancoGarantia: '',
  nroChequeGarantia: '',
}